backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "fitness"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
email-validator = "^2.1.1"
pyjwt = "^2.8.0"
pydantic-settings = "^2.2.1"
pymongo = "^4.6.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"
black = "^24.4.0"
isort = "^5.13.2"
coverage = "^7.4.4"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."

[tool.coverage.run]
source = "fitness"

[tool.coverage.report]
include_namespace_packages = true

[tool.mypy]
plugins = "pydantic.mypy"
